requirements.txt
Django==5.2.7
djangorestframework==3.15.2
celery==5.4.0
redis==5.2.0
psycopg2-binary==2.9.7

